feat: FilterUtility - Add filter util function
