+++ src/packages/util/index.ts
@@ -4,6 +4,7 @@ import { ObjectUtility } from "./ObjectUtility";
 import { DateUtility } from "./DateUtility";
 import { TestUtility } from "./TestUtility";
 import { WindowUtility } from "./WindowUtility";
+import { FilterUtility } from "./FilterUtility";
 import { UtilError } from "./UtilError";
 
 export * from "./types";
@@ -15,5 +16,6 @@ export {
   DateUtility,
   TestUtility,
   WindowUtility,
+  FilterUtility,
   UtilError,
 };
